rename and re-organize modules
  src/
  └── Cardano
      ├── Address
      │   ├── Derivation.hs
      │   └── Style
      │       ├── Byron.hs
      │       └── Icarus.hs
      ├── Address.hs
      ├── Codec
      │   └── Cbor.hs
      └── Mnemonic.hs
