+++ src/Cardano/Address/Derivation.hs
@@ -16,7 +16,7 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module Cardano.AddressDerivation
+module Cardano.Address.Derivation
     (
     -- * Overview
     -- $overview
@@ -25,20 +25,12 @@ module Cardano.AddressDerivation
       GenMasterKey (..)
     , HardDerivation (..)
     , SoftDerivation (..)
-    , PaymentAddress (..)
-
-    -- * Network Discrimination
-    , NetworkDiscriminant (..)
-    , ProtocolMagic (..)
-    , mainnetMagic
-    , testnetMagic
 
     -- * Helper types
     , Index (..)
     , Depth (..)
     , DerivationType (..)
     , AccountingStyle (..)
-    , Address (..)
     ) where
 
 import Prelude
@@ -49,10 +41,6 @@ import Control.DeepSeq
     ( NFData )
 import Data.ByteArray
     ( ScrubbedBytes )
-import Data.ByteString
-    ( ByteString )
-import Data.Proxy
-    ( Proxy (..) )
 import Data.String
     ( fromString )
 import Data.Typeable
@@ -63,8 +51,6 @@ import Fmt
     ( Buildable (..) )
 import GHC.Generics
     ( Generic )
-import GHC.TypeLits
-    ( KnownNat, Nat, natVal )
 
 -- $overview
 --
@@ -215,35 +201,3 @@ class GenMasterKey (key :: Depth -> * -> *) where
 
     -- | Generate a root key from a corresponding seed.
     genMasterKey :: GenMasterKeyFrom key -> ScrubbedBytes -> key 'RootK XPrv
-
-
--- | Address is mere wrapper around a ByteString and represents an encoded address.
-newtype Address = Address
-    { unAddress :: ByteString
-    } deriving (Show, Generic, Eq, Ord)
-
-instance NFData Address
-
--- | Available network options.
-data NetworkDiscriminant = Mainnet | Testnet Nat
-
--- | Magic constant associated to a given network
-newtype ProtocolMagic = ProtocolMagic { getProtocolMagic :: Word32 }
-    deriving (Generic, Show, Eq)
-
--- | Hard-coded protocol magic for the Byron MainNet
-mainnetMagic :: ProtocolMagic
-mainnetMagic =  ProtocolMagic 764824073
-
--- | Derive testnet magic from a type-level Nat
-testnetMagic :: forall pm. KnownNat pm => ProtocolMagic
-testnetMagic = ProtocolMagic $ fromIntegral $ natVal $ Proxy @pm
-
--- | Encoding of addresses for certain key types and backend targets.
-class PaymentAddress (network :: NetworkDiscriminant) key where
-    -- | Convert a public key to a payment 'Address' valid for the given
-    -- network discrimination.
-    --
-    -- Note that 'paymentAddress' is ambiguous and requires therefore a type
-    -- application.
-    paymentAddress :: key 'AddressK XPub -> Address
