+++ src/Cardano/Address/Style/Byron.hs
@@ -27,7 +27,7 @@
 -- see the "Cardano.Crypto.Wallet" module, and the implementation in
 -- <https://github.com/input-output-hk/cardano-crypto/blob/4590efa638397e952a51a8994b5543e4ea3c1ecd/cbits/encrypted_sign.c cardano-crypto>.
 
-module Cardano.AddressDerivation.Byron
+module Cardano.Address.Style.Byron
     ( -- * Types
       Byron(..)
 
@@ -41,16 +41,18 @@ module Cardano.AddressDerivation.Byron
 
 import Prelude
 
-import Cardano.AddressDerivation
+import Cardano.Address
     ( Address (..)
-    , Depth (..)
+    , NetworkDiscriminant (..)
+    , PaymentAddress (..)
+    , testnetMagic
+    )
+import Cardano.Address.Derivation
+    ( Depth (..)
     , DerivationType (..)
     , GenMasterKey (..)
     , HardDerivation (..)
     , Index (..)
-    , NetworkDiscriminant (..)
-    , PaymentAddress (..)
-    , testnetMagic
     )
 import Cardano.Crypto.Wallet
     ( DerivationScheme (DerivationScheme1)
