+++ test/Cardano/Address/Style/ByronSpec.hs
@@ -12,20 +12,20 @@
 
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
-module Cardano.AddressDerivation.ByronSpec
+module Cardano.Address.Style.ByronSpec
     ( spec
     ) where
 
 import Prelude
 
-import Cardano.AddressDerivation
+import Cardano.Address.Derivation
     ( AccountingStyle
     , Depth (..)
     , DerivationType (..)
     , HardDerivation (..)
     , Index (..)
     )
-import Cardano.AddressDerivation.Byron
+import Cardano.Address.Style.Byron
     ( Byron (..)
     , minSeedLengthBytes
     , mkByronKeyFromMasterKey
