+++ test/Cardano/Codec/CborSpec.hs
@@ -16,9 +16,9 @@ module Cardano.Codec.CborSpec
 
 import Prelude
 
-import Cardano.AddressDerivation
+import Cardano.Address.Derivation
     ( Depth (..), DerivationType (..), Index (..) )
-import Cardano.AddressDerivation.Byron
+import Cardano.Address.Style.Byron
     ( Byron (..), unsafeGenerateKeyFromSeed )
 import Cardano.Codec.Cbor
     ( decodeAddressDerivationPath
